+++ src/TestSuite/ReflectionTrait.php
@@ -12,7 +12,7 @@
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 
-namespace Tools;
+namespace Tools\TestSuite;
 
 use ReflectionClass;
 use ReflectionMethod;