+++ src/TestSuite/TestTrait.php
@@ -106,14 +106,14 @@ trait TestTrait
 
     /**
      * Asserts that a callable throws an exception
-     * @param string $expectedException Expected exception
      * @param callable $function A callable you want to test and that should
      *  raise the expected exception
+     * @param string $expectedException Expected exception
      * @param string|null $expectedMessage The expected message or `null`
      * @return void
      * @since 1.1.7
      */
-    protected static function assertException($expectedException, callable $function, $expectedMessage = null)
+    protected static function assertException(callable $function, $expectedException = Exception::class, $expectedMessage = null)
     {
         if ($expectedException !== Exception::class && !is_subclass_of($expectedException, Exception::class)) {
             self::fail(sprintf('Class `%s` does not exist or is not an exception', $expectedException));
@@ -181,40 +181,6 @@ trait TestTrait
     {
         self::assertFileExists($filename);
         self::assertContains(mime_content_type($filename), (array)$expectedMime, $message);
-    }
-
-    /**
-     * Asserts that a filename has some file permissions.
-     *
-     * If `$expectedPerms` is an array, asserts that the filename has at
-     *  least one of those values
-     * @param string|int|array $expectedPerms Expected permission values as a
-     *  four-chars string or octal value
-     * @param string $filename Filename
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     * @deprecated Use instead `assertFileIsReadable()`/`assertFileIsWritable()`/`assertDirectoryIsReadable()`/`assertDirectoryIsWritable()`
-     * @since 1.0.9
-     */
-    protected static function assertFilePerms($expectedPerms, $filename, $message = '')
-    {
-        deprecationWarning('Deprecated. Use instead `assertFileIsReadable()`/`assertFileIsWritable()`/`assertDirectoryIsReadable()`/`assertDirectoryIsWritable()`');
-
-        $isWritable = false;
-        foreach ((array)$expectedPerms as $perms) {
-            if (string_starts_with(fileperms_to_string($perms), '07')) {
-                $isWritable = true;
-
-                break;
-            }
-        }
-
-        $method = $isWritable ? 'assertFileIsWritable' : 'assertFileIsReadable';
-        if (is_dir($filename)) {
-            $method = $isWritable ? 'assertDirectoryIsWritable' : 'assertDirectoryIsReadable';
-        }
-        call_user_func_array(['parent', $method], [$filename, $message]);
     }
 
     /**